Correción lectura tiempos del json
Cambio de lectura del dt y el tRefresh por Int o Double y su consecuensia en las demás clases.
Corrrección de cálculo en los tiempos de ResultadosSimulación.
Adición de los nombres al README

+++ src/com/unalmed/vehTraffic/dimension/MovimientoUniforme.scala
@@ -16,7 +16,7 @@ trait MovimientoUniforme{
   val ruta : Queue[Via]
   val intersecciones : Queue[Interseccion]
   
-  def cambioPosicion(dt : Int) = {
+  def cambioPosicion(dt : Double) = {
     var tiempo = dt
     if(!intersecciones.isEmpty){
       while(tiempo !=0){
