+++ src/com/unalmed/vehTraffic/vehiculo/Movil.scala
@@ -9,6 +9,6 @@ abstract class Movil(protected var _posicion : Punto, protected var _velocidad :
   def posicion = _posicion
   def velocidad_=(vel:Velocidad):Unit
   def posicion_=(pos:Punto):Unit
-  def cambioPosicion(dt : Int)
+  def cambioPosicion(dt : Double)
   def angulo = velocidad.direccion.valor
 }
